# feat(site-builder): parallelize calls to the walrus cli