--- a/site-builder/Cargo.toml
+++ b/site-builder/Cargo.toml
@@ -27,7 +27,7 @@ sui-keys = { git = "https://github.com/MystenLabs/sui", tag = "testnet-v1.34.2" 
 sui-sdk = { git = "https://github.com/MystenLabs/sui", tag = "testnet-v1.34.2" }
 sui-types = { git = "https://github.com/MystenLabs/sui", tag = "testnet-v1.34.2" }
 thiserror = "1.0.61"
-tokio = { version = "1.38.0", features = ["rt", "rt-multi-thread", "macros"] }
+tokio = { version = "1.38.0", features = ["rt", "rt-multi-thread", "macros", "process"] }
 toml = "0.8.14"
 tracing = "0.1.40"
 tracing-subscriber = "0.3.18"