@giac-mysten 2024-10-10
@chariskms this PR touches also the `Unchanged` resource operation. I realized that trying to re-upload unchanged resources as a way to achieve #35 is fine for small sites, but increases latency massively for big ones. One way to get around this problem is to go through the `blob-info` endpoint (instead of `store`), because it can be parallelized. 

Therefore, a solution could look like:
* find the set of unchanged resources (already implemented)
* check the blob status by parallel calls to the `blob-status` endpoint
* for the calls that return `nonexistent`, upload the blob again

however, one issue with this is that each of the calls requires a query to the full node. When done in parallel, these will quickly reach the rate limit for the public full node and the site builder will fail. 

So i'm not too sure what the best way forward is here. Thoughts?